requirements.txt
streamlit>=1.28
numpy>=1.24
plotly>=5.17
matplotlib>=3.7
scipy>=1.11
requests>=2.31.0
Pillow>=10.0.0
pandas==2.0.3